+++ app/models/Game.scala
@@ -5,6 +5,6 @@ import org.squeryl.KeyedEntity
 class Game(
             val name: String,
             val code: String,
-          ) extends KeyedEntity[Int] {
-  val id: Int = 0
+          ) extends KeyedEntity[Long] {
+  val id: Long = 0
 }
