+++ app/v1/transfer/TransferRouter.scala
@@ -10,7 +10,7 @@ import play.api.routing.sird._
 class TransferRouter @Inject()(controller: TransferController) extends SimpleRouter {
   val prefix = "/v1/transfers"
 
-  def link(id: Int): String = {
+  def link(id: Long): String = {
     import com.netaporter.uri.dsl._
     val url = prefix / id.toString
     url.toString()
